git-tfs-id: [http://192.168.56.101:8080/tfs/DefaultCollection]$/TikiEngine 2.0;C1171

+++ src/Game/TikiSteering.cpp
@@ -89,8 +89,8 @@ namespace TikiEngine
 		{
 			Vector2 separationForce = Vector2::Zero;
 
-			tikiBot->GetGameState()->GetScene()->SceneGraph.DoWithinRange(tikiBot->Pos3D(), tikiBot->BRadius() + 1, [&](GameObject* go) 
-			{ 
+			tikiBot->GetGameState()->GetScene()->SceneGraph.DoWithinRange(tikiBot->Pos3D(), (float)(tikiBot->BRadius() + 1.0), [&](GameObject* go) 
+			{
 
  				if (go != 0)
  				{
